requirements.txt
streamlit
python-docx
pandas
openpyxl
graphviz